backend/requirements.txt
fastapi
uvicorn
pydantic
groq
python-dotenv
authlib
python-jose[cryptography]
